support rms_norm for torch_dipu and torch_npu
