+++ tests/easyllm/test_rms_norm_dipu.py
